+++ db/dungeons/fh.lua
@@ -32,7 +32,7 @@ FH.Entrances = {
     },
 }
 FH.Bosses = {
-    [1] = { -- NOTE: Missing abilities, Missing Loot
+    [1] = { -- NOTE: Missing Loot
         Id = "tansha",
         Portrait = IMP .. IMPD .. instanceMediaAcronym .. "tansha" .. IMPP,
         Frame = IMP .. IMPD .. instanceMediaAcronym .. "tansha" .. IMPF,
@@ -49,11 +49,61 @@ FH.Bosses = {
                 Quality = IQ.Uncommon,
                 ItemSlot = IS.Back,
                 ItemType = IT.Cloth,
-                DropChance = 25,
+                DropChance = 33,
+                LevelRequirement = 13,
+            },
+            [2] = {
+                Id = "184",
+                Icon = "INV_ThrowingAxe_01",
+                Name = IJ_LOOT_NAME_184,
+                Quality = IQ.Uncommon,
+                ItemSlot = IS.TwoHand,
+                ItemType = IT.Axe,
+                DropChance = 33,
+                LevelRequirement = 13,
+            },
+            [3] = {
+                Id = "197",
+                Icon = "INV_Chest_Leather_04",
+                Name = IJ_LOOT_NAME_197,
+                Quality = IQ.Uncommon,
+                ItemSlot = IS.Chest,
+                ItemType = IT.Leather,
+                DropChance = 33,
+                LevelRequirement = 13,
+            },
+        },
+        Abilities = {
+            [1] = {
+                Id = "tansha",
+                Icon = "ability_mount_whitetiger",
+                Name = IJ_DB_DUNGEON_FH_BOSS_tansha_ABILITY_NAME_tansha,
+                SubAbilities = {
+                    [1] = {
+                        Id = "tanshaintimidatingshout",
+                        Icon = "ability_golemthunderclap",
+                        Name = IJ_DB_DUNGEON_FH_BOSS_tansha_ABILITY_NAME_tanshaintimidatingshout,
+                        Effect = IJ_DB_DUNGEON_FH_BOSS_tansha_ABILITY_EFFECT_tanshaintimidatingshout,
+                        Flags = { IMAF.Important, },
+                    },
+                },
+            },
+            [2] = {
+                Id = "oboka",
+                Icon = "inv_ammo_arrow_02",
+                Name = IJ_DB_DUNGEON_FH_BOSS_tansha_ABILITY_NAME_oboka,
+                SubAbilities = {
+                    [1] = {
+                        Id = "obokamendpet",
+                        Icon = "ability_hunter_mendpet",
+                        Name = IJ_DB_DUNGEON_FH_BOSS_tansha_ABILITY_NAME_obokamendpet,
+                        Effect = IJ_DB_DUNGEON_FH_BOSS_tansha_ABILITY_EFFECT_obokamendpet,
+                    },
+                },
             },
         },
     },
-    [2] = { -- NOTE: Missing abilities, Missing Loot
+    [2] = { -- NOTE: Missing Loot
         Id = "ubukaz",
         Portrait = IMP .. IMPD .. instanceMediaAcronym .. "ubukaz" .. IMPP,
         Frame = IMP .. IMPD .. instanceMediaAcronym .. "ubukaz" .. IMPF,
@@ -62,6 +112,15 @@ FH.Bosses = {
         MapCoordinateX = "52.4",
         MapCoordinateY = "39.4",
         MapId = "1",
+        Abilities = {
+            [1] = {
+                Id = "enrage",
+                Icon = "spell_shadow_unholyfrenzy",
+                Name = IJ_DB_DUNGEON_FH_BOSS_ubukaz_ABILITY_NAME_enrage,
+                Effect = IJ_DB_DUNGEON_FH_BOSS_ubukaz_ABILITY_EFFECT_enrage,
+                Flags = { IMAF.Enrage, },
+            },
+        },
         Loot = {
             [1] = {
                 Id = "150",
@@ -70,11 +129,12 @@ FH.Bosses = {
                 Quality = IQ.Uncommon,
                 ItemSlot = IS.Hands,
                 ItemType = IT.Mail,
-                DropChance = 25,
+                DropChance = 100,
+                LevelRequirement = 13,
             },
         },
     },
-    [3] = { -- NOTE: Missing abilities, Missing Loot
+    [3] = { -- NOTE: Missing Loot
         Id = "kanza",
         Portrait = IMP .. IMPD .. instanceMediaAcronym .. "kanza" .. IMPP,
         Frame = IMP .. IMPD .. instanceMediaAcronym .. "kanza" .. IMPF,
@@ -91,12 +151,44 @@ FH.Bosses = {
                 Quality = IQ.Uncommon,
                 ItemSlot = IS.OneHand,
                 ItemType = IT.Dagger,
-                DropChance = 25,
+                DropChance = 100,
+                LevelRequirement = 13,
             },
         },
-        -- NOTE: Has blizzard ability
+        Abilities = {
+            [1] = {
+                Id = "kanzaadds",
+                Icon = "inv_misc_head_troll_01",
+                Name = IJ_DB_DUNGEON_FH_BOSS_kanza_ABILITY_NAME_kanzaadds,
+                Effect = IJ_DB_DUNGEON_FH_BOSS_kanza_ABILITY_EFFECT_kanzaadds,
+                Flags = { IMAF.DPS, },
+                SubAbilities = {
+                    [1] = {
+                        Id = "kanzaaddsfrostbolt",
+                        Icon = "spell_frost_frostbolt02",
+                        Name = IJ_DB_DUNGEON_FH_BOSS_kanza_ABILITY_NAME_kanzaaddsfrostbolt,
+                        Effect = IJ_DB_DUNGEON_FH_BOSS_kanza_ABILITY_EFFECT_kanzaaddsfrostbolt,
+                        Flags = { IMAF.Interruptible, },
+                    },
+                },
+            },
+            [2] = {
+                Id = "kanzablizzard",
+                Icon = "spell_frost_icestorm",
+                Name = IJ_DB_DUNGEON_FH_BOSS_kanza_ABILITY_NAME_kanzablizzard,
+                Effect = IJ_DB_DUNGEON_FH_BOSS_kanza_ABILITY_EFFECT_kanzablizzard,
+                Flags = { IMAF.Interruptible, },
+            },
+            [3] = {
+                Id = "kanzafrostbolt",
+                Icon = "spell_frost_frostbolt02",
+                Name = IJ_DB_DUNGEON_FH_BOSS_kanza_ABILITY_NAME_kanzafrostbolt,
+                Effect = IJ_DB_DUNGEON_FH_BOSS_kanza_ABILITY_EFFECT_kanzafrostbolt,
+                Flags = { IMAF.Interruptible, },
+            },
+        },
     },
-    [4] = { -- NOTE: Missing abilities, Missing Loot
+    [4] = { -- NOTE: Missing Loot
         Id = "hailar",
         Portrait = IMP .. IMPD .. instanceMediaAcronym .. "hailar" .. IMPP,
         Frame = IMP .. IMPD .. instanceMediaAcronym .. "hailar" .. IMPF,
@@ -105,6 +197,29 @@ FH.Bosses = {
         MapCoordinateX = "30.5",
         MapCoordinateY = "86.1",
         MapId = "1",
+        Abilities = {
+            [1] = {
+                Id = "hailaradds",
+                Icon = "inv_misc_head_troll_01",
+                Name = IJ_DB_DUNGEON_FH_BOSS_hailar_ABILITY_NAME_hailaradds,
+                Effect = IJ_DB_DUNGEON_FH_BOSS_hailar_ABILITY_EFFECT_hailaradds,
+                Flags = { IMAF.DPS, IMAF.Important, },
+            },
+            [2] = {
+                Id = "16803",
+                Icon = "spell_frost_frostnova",
+                Name = IJ_DB_DUNGEON_FH_BOSS_hailar_ABILITY_NAME_16803,
+                Effect = IJ_DB_DUNGEON_FH_BOSS_hailar_ABILITY_EFFECT_16803,
+                Flags = { IMAF.Deadly, },
+            },
+            [3] = {
+                Id = "hailarfrostbolt",
+                Icon = "spell_frost_frostbolt02",
+                Name = IJ_DB_DUNGEON_FH_BOSS_hailar_ABILITY_NAME_hailarfrostbolt,
+                Effect = IJ_DB_DUNGEON_FH_BOSS_hailar_ABILITY_EFFECT_hailarfrostbolt,
+                Flags = { IMAF.Interruptible, },
+            },
+        },
         Loot = {
             [1] = {
                 Id = "110",
@@ -113,7 +228,27 @@ FH.Bosses = {
                 Quality = IQ.Rare,
                 ItemSlot = IS.MainHand,
                 ItemType = IT.Sword,
-                DropChance = 25,
+                DropChance = 33,
+                LevelRequirement = 13,
+            },
+            [2] = {
+                Id = "126",
+                Icon = "INV_Jewelry_Ring_28",
+                Name = IJ_LOOT_NAME_126,
+                Quality = IQ.Rare,
+                ItemSlot = IS.Finger,
+                DropChance = 33,
+                LevelRequirement = 13,
+            },
+            [3] = {
+                Id = "116",
+                Icon = "INV_Belt_03",
+                Name = IJ_LOOT_NAME_116,
+                Quality = IQ.Rare,
+                ItemSlot = IS.Waist,
+                ItemType = IT.Leather,
+                DropChance = 33,
+                LevelRequirement = 13,
             },
         },
     },