+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All changes to this project will be documented in this file.
 
+## [1.0.3] - 2026-03-21
+
+### Added
+
+- Added most loot from Frostmane Hollow
+- Added all abilities from Frostmane Hollow
+- Added most loot from Windhorn Canyon
+- Added some abilities from Windhorn Canyon
+- Added all (?) loot from Broodcommander Axelus
+
 ## [1.0.2] - 2026-03-21
 
 ### Fixed